Реализованы чаты, изменение аватарки, вывод заявок в друзья их принятие, а так же приватные чаты

+++ backend/src/freind/freind.service.ts
@@ -7,6 +7,7 @@ import { User } from 'src/user/user.model';
 import { Freind } from './freind.model';
 import { GetRequestDto } from './dto/get-request-dto';
 import { AcceptRequestDto } from './dto/accept-request-dto';
+import { Op } from 'sequelize';
 
 @Injectable()
 export class FreindService {
@@ -24,6 +25,24 @@ export class FreindService {
         freindId: createRequestDto.freindId,
       },
     });
+
+    const checkIsFreind = await this.freindRepository.findOne({
+      where: {
+        [Op.or]: [
+          { userId: createRequestDto.userId, freindId: createRequestDto.freindId },
+          { userId: createRequestDto.freindId, freindId: createRequestDto.userId }
+        ]
+      }
+    })
+    
+
+    if (checkIsFreind) {
+      throw new HttpException(
+        'Пользователь уже у вас в друзьях',
+        HttpStatus.CONFLICT,
+      );
+    }
+
     if (requestCheck) {
       throw new HttpException(
         'Заявка на добавление в друзья уже отправлена',
@@ -80,6 +99,19 @@ export class FreindService {
   }
 
   async acceptRequest(acceptRequestDto: AcceptRequestDto) {
+    const checkIsConflict = await this.freindRepository.findOne({
+      where: {
+        [Op.or]: [
+          { userId: acceptRequestDto.userId, freindId: acceptRequestDto.freindId },
+          { userId: acceptRequestDto.freindId, freindId: acceptRequestDto.userId }
+        ]
+      }
+    })
+
+    if(checkIsConflict) {
+      throw new HttpException('Заявка на добавление в друзья уже создана', HttpStatus.CONFLICT);
+    }
+
     await this.freindRequestRepository.destroy({
       where: { id: acceptRequestDto.requestId },
     });
